Player object moves and is network aware
Movement is just forward and back with w and d, with rotation with a and s
right now until I can get the camera attached to the face of the player.

Networking is working. Building the project and starting a host, then
starting the game in Unity and starting a client creates two players, which
move on both instances (using localhost).