Bug fix in how health bar is updated. Player now respawns on death.
Health bar was not displaying correctly. Now the playerHealth script gets
the sizeDelta.x of the health bar on start and uses that value to calculate
current health as a percent of the total size of the bar to display health
regardless of the size of the health bar.

Player respawns immediately, for now. Respawning only entails resetting
health to max and resetting position to 0.