Movement/looking improvements.
The "gun" now moves up/down with the camera. The mouse angle is clamped
between -80 and 80 degrees.

Now that the camera is working correctly, A and D move the player side to
side instead of rotating. Adding a rigidbody and getting the bullets to
interact with it correctly is high on the priority list